requirements.txt
ultralytics
opencv-python
pandas 
pyyaml